Fixes #27: Disable nullable reference types for options class
There's a bug in the commandline parser library that triggers when an option class has a custom usage function
and that function has more than one attribute.
When a class has nullable reference type properties, it's members are all tagged with the System.Runtime.CompilerServices.NullableAttribute attribute.
Disabling NRTs in just the Options.cs file fixes the bug while still allowing us to use NRTs elsewhere.

A bug has been submitted upstream
https://github.com/commandlineparser/commandline/issues/714

As well as a pull request that fixes the issue
https://github.com/commandlineparser/commandline/pull/715

The compiler directive can be removed once it's accepted and released.